+++ jeap-deploymentlog-docgen/src/main/java/ch/admin/bit/jeap/deploymentlog/docgen/model/DeploymentHistoryOverviewPageDto.java
@@ -14,8 +14,12 @@ public class DeploymentHistoryOverviewPageDto {
     String deploymentHistoryOverviewMinStartedAt;
     List<DeploymentDto> deployments;
 
-    public String getPageTitle() {
+    public static String pageTitle(String environmentName) {
         return "Deployment History Overview " + environmentName;
+    }
+
+    public String getPageTitle() {
+        return pageTitle(environmentName);
     }
 
 }